Move and export errorInterceptor
