+++ src/lib/seam/connect/index.ts
@@ -1,3 +1,4 @@
+export * from './error-interceptor.js'
 export * from './options.js'
 export * from './routes/index.js'
 export * from './seam-http.js'
